+++ tests/unit/provenance/test_provenance.py
@@ -8,9 +8,9 @@ import json
 import re
 import secrets
 
-from starfish.metadata.provenance import create_publish, create_invoke
+from starfish.provenance.provenance import create_publish, create_invoke
 from starfish.asset import DataAsset, OperationAsset
-from starfish.metadata import METADATA_PROVENANCE
+from starfish.metadata import METADATA_PROVENANCE_NAME
 
 OPERATION_METADATA = {
     'name': 'operation',
@@ -130,7 +130,7 @@ def test_assign_provenance_to_data_asset():
     agent_id = secrets.token_hex(32)
     test_data = secrets.token_hex(120)
     metadata = {
-        METADATA_PROVENANCE: create_publish(agent_id)
+        METADATA_PROVENANCE_NAME: create_publish(agent_id)
     }
     asset = DataAsset.create('test data asset', test_data, metadata)
     assert(asset)
@@ -155,7 +155,7 @@ def test_assign_provenance_to_operation_asset():
         }
     )
     metadata = OPERATION_METADATA
-    metadata[METADATA_PROVENANCE] = create_invoke(activity_id, agent_id, asset_list, inputs_text, outputs_text)
+    metadata[METADATA_PROVENANCE_NAME] = create_invoke(activity_id, agent_id, asset_list, inputs_text, outputs_text)
     asset = OperationAsset(OPERATION_METADATA)
     assert(asset)
     assert(isinstance(asset, OperationAsset))
